information_extraction/requirements.txt
openai
pdf2image
Pillow
python-dotenv
